# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Service Portal API",
    "version": "1.0.0",
...
  },
  "paths": {
...
    "/api/v1/service-requests/{request_id}/status-changes": {
      "get": {
        "tags": [
          "status-changes"
        ],
        "summary": "List a request's status history",
        "description": "This request's transitions, oldest first (SC-1 to SC-3, XC-10, XC-11).\n\nNo role branch: SC-1 gives the history to whoever can see the request, and\nthe parent dependency has already decided that. Unlike comments, there is no\nper-row audience — a transition is a fact about the request, not a note\nabout it.\n\nEvery row returned is a row that exists (SC-3). That is not enforced by\nanything here so much as by there being nothing here that could invent one:\nthe un-reached statuses a stepper draws hollow are absent from this query's\ntable, and design.md §5 puts that merge in the frontend deliberately.",
        "operationId": "list_status_changes_api_v1_service_requests__request_id__status_changes_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StatusChangeOut_"
                }
              }
            }
          },
          "422": {
            "description": "Returned when one or more supplied fields are invalid (XC-4).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the `access_token` cookie is absent, expired, or no longer names an active user — no distinction is drawn between those cases (XC-5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Returned when no service request with that id is visible to the caller — a malformed id, an id matching no row, and an id belonging to someone else are answered identically by design (XC-7, SR-13).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessTokenCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "status-changes"
        ],
        "summary": "Transition a request to a new status",
        "description": "Transition a request, as an admin (SC-4 to SC-7).\n\n``changed_by_id`` is the authenticated admin and is never read from the body\n(XC-8) — the schema has no field for it, so a body supplying one produces\nexactly the same row as one that doesn't.",
        "operationId": "create_status_change_api_v1_service_requests__request_id__status_changes_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Requested-With",
              "default": "XMLHttpRequest"
            },
            "description": "CSRF guard (XC-9). Any non-empty value. Enforced by middleware before routing, so a request without it is answered `403` without its body or its cookies being read — which is why it is absent from this operation's handler signature."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChangeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusChangeOut"
                }
              }
            }
          },
          "422": {
            "description": "Returned when one or more supplied fields are invalid (XC-4).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the `access_token` cookie is absent, expired, or no longer names an active user — no distinction is drawn between those cases (XC-5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Returned when the `X-Requested-With` header is missing, which middleware refuses before routing (XC-9); or when the caller's role is below the one this operation requires (XC-6).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Returned when no service request with that id is visible to the caller — a malformed id, an id matching no row, and an id belonging to someone else are answered identically by design (XC-7, SR-13).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessTokenCookie": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Page_StatusChangeOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StatusChangeOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[StatusChangeOut]"
      },
...
      "StatusChangeCreate": {
        "properties": {
          "status_id": {
            "type": "string",
            "format": "uuid",
            "title": "Status Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "status_id"
        ],
        "title": "StatusChangeCreate",
        "description": "``POST /service-requests/{id}/status-changes`` body (SC-5 to SC-7)."
      },
      "StatusChangeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/StatusOut"
          },
          "changed_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "changed_at": {
            "type": "string",
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "changed_by",
          "note",
          "changed_at"
        ],
        "title": "StatusChangeOut",
        "description": "One transition, as design.md §5 writes it.\n\nEvery field is of a row that actually happened. SC-3 forbids synthesising a\nplaceholder for a status not yet reached, so there is no \"pending\" variant\nof this shape and no nullable ``changed_at`` — the frontend's stepper gets\nits un-reached steps by diffing this list against ``GET /statuses``, not by\nreceiving hollow rows from here."
      },
...
      "StatusOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "is_terminal": {
            "type": "boolean",
            "title": "Is Terminal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "sort_order",
          "is_terminal"
        ],
        "title": "StatusOut",
        "description": "One row of the ``statuses`` reference table (design.md §1)."
      },
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "role"
        ],
        "title": "UserSummary",
        "description": "A user reference embedded in another resource (design.md §1)."
      },
      "ErrorDetail": {
        "title": "ErrorDetail",
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "title": "Code",
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_ERROR",
              "INTERNAL_ERROR"
            ],
            "description": "Machine-readable error class. Closed set."
          },
          "message": {
            "title": "Message",
            "type": "string",
            "description": "Human-readable summary, safe to show a user. Never carries server internals."
          },
          "fields": {
            "title": "Fields",
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Per-field messages, keyed by field name. Present only on `VALIDATION_ERROR`."
          }
        }
      },
      "ErrorEnvelope": {
        "title": "ErrorEnvelope",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "description": "The one shape every error response in this API uses."
      }
    },
    "securitySchemes": {
      "accessTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "Session access token. Set by `POST /api/v1/auth/login` and `POST /api/v1/auth/refresh` as an `httpOnly`, `Secure`, `SameSite=Lax` cookie with a 1 hour lifetime. Sent automatically by the browser; not readable from JavaScript and not settable by a client."
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "Opaque refresh token, 30 day lifetime, same cookie attributes as `access_token`. Read by `POST /api/v1/auth/refresh` only, which rotates it; replaying an already-rotated token revokes the whole family (AUTH-11)."
      }
    }
  }
}